+++ source/extensions/omni.isaac.core/omni/isaac/core/tasks/stacking.py
@@ -10,7 +10,6 @@ from abc import abstractmethod, ABC
 from omni.isaac.core.tasks import BaseTask
 from omni.isaac.core.scenes.scene import Scene
 from omni.isaac.core.objects import DynamicCuboid
-from omni.isaac.core.articulations import ArticulationGripper
 import numpy as np
 from omni.isaac.core.utils.prims import is_prim_path_valid
 from omni.isaac.core.utils.stage import get_stage_units
@@ -174,8 +173,10 @@ class Stacking(ABC, BaseTask):
     def post_reset(self) -> None:
         """[summary]
         """
-        if isinstance(self._robot.gripper, ArticulationGripper):
-            self._robot.gripper.set_positions(self._robot.gripper.open_position)
+        from omni.isaac.manipulators.grippers.parallel_gripper import ParallelGripper
+
+        if isinstance(self._robot.gripper, ParallelGripper):
+            self._robot.gripper.set_joint_positions(self._robot.gripper.joint_opened_positions)
         return
 
     def get_cube_names(self) -> List[str]: